interface_collection module was causing weird analysis issues. Rename to interface. Update Chumsky to 0.8.0. Change delimited_by to use combinators.
