+++ crates/til_parser/Cargo.toml
@@ -10,7 +10,7 @@ description = "A grammar and parser for the Tydi Intermediate representation Lan
 [dependencies]
 log = "0.4.14"
 ariadne = "0.1.3"
-chumsky = "0.7.0"
+chumsky = "0.8.0"
 indexmap = "1.7.0"
 tydi-common = { path = "../common" }
 tydi-intern = { path = "../intern" }
