+++ crates/til_query/src/ir/intern_self.rs
@@ -1,7 +1,7 @@
 use tydi_intern::Id;
 
 use super::{
-    project::{interface_collection::InterfaceCollection, namespace::Namespace},
+    project::{interface::InterfaceCollection, namespace::Namespace},
     traits::InternSelf,
     Implementation, Interface, Ir, LogicalType, Stream, Streamlet,
 };
