+++ crates/til_query/src/ir/project/mod.rs
@@ -17,7 +17,7 @@ use super::{
     Ir,
 };
 
-pub mod interface_collection;
+pub mod interface;
 pub mod namespace;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
